refactor: update imports
Updates several imports. Removes generic type parameters that are no
longer in use.
