+++ build.gradle
@@ -38,14 +38,14 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
     implementation 'org.springframework.security:spring-security-web'
 
-    implementation 'no.fintlabs:fint-core-infra-models:1.1.2'
-    implementation 'no.fintlabs:fint-core-adapter-common:0.1.5-rc-2'
+    implementation 'no.fintlabs:fint-core-infra-models:2.1.1'
+    implementation 'no.fintlabs:fint-core-adapter-common:0.1.6-rc-7'
     implementation "no.fint:fint-utdanning-resource-model-java:${apiVersion}"
     implementation 'no.fint:fint-model-resource:0.4.1'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    compileOnly 'org.projectlombok:lombok'
-    annotationProcessor 'org.projectlombok:lombok'
+    compileOnly 'org.projectlombok:lombok:1.18.42'
+    annotationProcessor 'org.projectlombok:lombok:1.18.42'
     testImplementation 'org.wiremock:wiremock-standalone:3.13.1'
     testImplementation 'org.junit.jupiter:junit-jupiter:5.10.2'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
