+++ src/main/java/no/fintlabs/model/larling/LarlingRepository.java
@@ -2,7 +2,7 @@ package no.fintlabs.model.larling;
 
 import no.fint.model.resource.utdanning.larling.LarlingResource;
 import no.fintlabs.adapter.events.WriteableResourceRepository;
-import no.fintlabs.adapter.models.RequestFintEvent;
+import no.fintlabs.adapter.models.event.RequestFintEvent;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
