+++ src/main/java/no/fintlabs/model/larling/LarlingSubscriber.java
@@ -4,7 +4,7 @@ import no.fint.model.resource.utdanning.larling.LarlingResource;
 import no.fintlabs.adapter.config.AdapterProperties;
 import no.fintlabs.adapter.datasync.ResourceSubscriber;
 import no.fintlabs.adapter.models.AdapterCapability;
-import no.fintlabs.adapter.models.SyncPageEntry;
+import no.fintlabs.adapter.models.sync.SyncPageEntry;
 import no.fintlabs.adapter.validator.ValidatorService;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -12,7 +12,7 @@ import org.springframework.web.reactive.function.client.WebClient;
 @Service
 public class LarlingSubscriber extends ResourceSubscriber<LarlingResource, LarlingPublisher> {
 
-    protected LarlingSubscriber(WebClient webClient, AdapterProperties props, LarlingPublisher publisher, ValidatorService<LarlingResource> validatorService) {
+    protected LarlingSubscriber(WebClient webClient, AdapterProperties props, LarlingPublisher publisher, ValidatorService validatorService) {
         super(webClient, props, publisher, validatorService);
     }
 
@@ -22,7 +22,7 @@ public class LarlingSubscriber extends ResourceSubscriber<LarlingResource, Larli
     }
 
     @Override
-    protected SyncPageEntry<LarlingResource> createSyncPageEntry(LarlingResource resource) {
+    protected SyncPageEntry createSyncPageEntry(LarlingResource resource) {
         return SyncPageEntry.of(resource.getSystemId().getIdentifikatorverdi(), resource);
     }
 }
