+++ src/main/java/no/fintlabs/model/person/PersonSubscriber.java
@@ -4,7 +4,7 @@ import no.fint.model.resource.felles.PersonResource;
 import no.fintlabs.adapter.config.AdapterProperties;
 import no.fintlabs.adapter.datasync.ResourceSubscriber;
 import no.fintlabs.adapter.models.AdapterCapability;
-import no.fintlabs.adapter.models.SyncPageEntry;
+import no.fintlabs.adapter.models.sync.SyncPageEntry;
 import no.fintlabs.adapter.validator.ValidatorService;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -12,7 +12,7 @@ import org.springframework.web.reactive.function.client.WebClient;
 @Service
 public class PersonSubscriber extends ResourceSubscriber<PersonResource, PersonPublisher> {
 
-    protected PersonSubscriber(WebClient webClient, AdapterProperties props, PersonPublisher publisher, ValidatorService<PersonResource> validatorService) {
+    protected PersonSubscriber(WebClient webClient, AdapterProperties props, PersonPublisher publisher, ValidatorService validatorService) {
         super(webClient, props, publisher, validatorService);
     }
 
@@ -22,7 +22,7 @@ public class PersonSubscriber extends ResourceSubscriber<PersonResource, PersonP
     }
 
     @Override
-    protected SyncPageEntry<PersonResource> createSyncPageEntry(PersonResource resource) {
+    protected SyncPageEntry createSyncPageEntry(PersonResource resource) {
         return SyncPageEntry.of(resource.getFodselsnummer().getIdentifikatorverdi(), resource);
     }
 }
