+++ src/main/java/no/fintlabs/model/virksomhet/VirksomhetRepository.java
@@ -2,7 +2,7 @@ package no.fintlabs.model.virksomhet;
 
 import no.fint.model.resource.felles.VirksomhetResource;
 import no.fintlabs.adapter.events.WriteableResourceRepository;
-import no.fintlabs.adapter.models.RequestFintEvent;
+import no.fintlabs.adapter.models.event.RequestFintEvent;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
